JavaScript Quiz Game
A JavaScript quiz game is an interactive learning tool featuring dynamic questions, instant feedback, and a user-friendly interface. It's educational, entertaining, and offers coding challenges to developers. Customizable and community-friendly, it promotes continuous learning and engagement.
These are Key points of the projects are:

User-Friendly Interface: HTML and CSS are employed to create a user-friendly and visually appealing interface. The design is often responsive, ensuring a seamless experience across different devices.
Score Tracking:  It keeps track of users scores and displays them in real-time. as the game records user answers and validates them against correct answers.
End Game Evaluation: Calculates and displays the final score at the end of the quiz.